[Fix #146706545] fix irregular styling
