+++ package.json
@@ -7,7 +7,7 @@
     "test": "NODE_ENV=test nyc mocha --compilers js:babel-core/register ./**/*.spec.js",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "test-watch": "nodemon --exec 'npm test'",
-    "start": "babel-node server.js"
+    "start": "nodemon --exec 'babel-node server'"
   },
   "nyc": {
     "require": [
